+++ modules/sor/sorV2/lib/utils/helpers.ts
@@ -69,53 +69,3 @@ export function getOutputAmount(paths: PathWithAmount[]): TokenAmount {
     const amounts = paths.map((path) => path.outputAmount);
     return amounts.reduce((a, b) => a.add(b));
 }
-
-export function getSwaps(paths: PathWithAmount[], swapKind: SwapKind) {
-    const isBatchSwap = paths.length > 1 || paths[0].pools.length > 1;
-    const assets = [...new Set(paths.flatMap((p) => p.tokens).map((t) => t.address))];
-
-    let swaps: BatchSwapStep[] | SingleSwap;
-    if (isBatchSwap) {
-        swaps = [] as BatchSwapStep[];
-        if (swapKind === SwapKind.GivenIn) {
-            paths.map((p) => {
-                p.pools.map((pool, i) => {
-                    (swaps as BatchSwapStep[]).push({
-                        poolId: pool.id,
-                        assetInIndex: BigInt(assets.indexOf(p.tokens[i].address)),
-                        assetOutIndex: BigInt(assets.indexOf(p.tokens[i + 1].address)),
-                        amount: i === 0 ? p.inputAmount.amount : 0n,
-                        userData: DEFAULT_USERDATA,
-                    });
-                });
-            });
-        } else {
-            paths.map((p) => {
-                // Vault expects given out swaps to be in reverse order
-                const reversedPools = [...p.pools].reverse();
-                const reversedTokens = [...p.tokens].reverse();
-                reversedPools.map((pool, i) => {
-                    (swaps as BatchSwapStep[]).push({
-                        poolId: pool.id,
-                        assetInIndex: BigInt(assets.indexOf(reversedTokens[i + 1].address)),
-                        assetOutIndex: BigInt(assets.indexOf(reversedTokens[i].address)),
-                        amount: i === 0 ? p.outputAmount.amount : 0n,
-                        userData: DEFAULT_USERDATA,
-                    });
-                });
-            });
-        }
-    } else {
-        const path = paths[0];
-        const pool = path.pools[0];
-        swaps = {
-            poolId: pool.id,
-            kind: swapKind,
-            assetIn: path.tokens[0].address,
-            assetOut: path.tokens[1].address,
-            amount: path.swapAmount.amount,
-            userData: DEFAULT_USERDATA,
-        } as SingleSwap;
-    }
-    return swaps;
-}
